Move `NumericType` and `PixelType` to a new file
